+++ svd-rs/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 
 - `default_value` for `EnumeratedValues`
+- move `EnumeratedValue` check to `strict` mode and fix it
 
 ## [v0.14.4] - 2023-11-15